sensors/requirements.txt
pyserial==3.5
firebase-admin==7.1.0
python-dotenv==1.0.0
flask==3.0.0
